Add files via upload
7805

I am going to show the working of a constant current battery charger circuit. They use a few components, 7805 and some transistors. We often use 7805 in DC regulator, 5V 1A. A lot of digital circuits use them as a source. It can keep a stable voltage.

 

In the datasheet, 7805 has many helpful. One thing is we can use them as constant current regulation. It is so easiest and more convenient than using transistor circuits.

Because of a one-man show with a few parts. So, become easily the battery charger circuit.

Imagine you use a digital multimeter, right? Because it is highly accurate and using easy. It uses a 9-volts battery as the main energy. When we often apply it use it very consumption. It runs out of energy. You need to change to a new one.

What can you do it?

Use a rechargeable battery of 9 volts (8.4V). It is rechargeable over 200 times so worth it. But we need the right charging way.

Let me show you a nice 9V Ni-MH battery charger.

Of course, when you finished reading. You can apply it to charge any battery like 7.2V, 12V battery, and more. As your ideas.

Basic current constant or current regulator using 7805

In the datasheet will see that a resistor-R1 passes current from a pin output of IC to load. Then, it gets the current output to pin ground, too.

The circuit inside 7805 can keep the output current in solid status. Even we change any input voltage. But do not forget it run well over 5V input.

What is more? Want to know how to find R1?

If you say yes, read below…

7805 Current Regulator Calculator

You can find the current constant or current regulator calculator, below

IO = 5V / R1

Defined:

IO: the current constant output

5V: Voltage of IC-7805

R1: Resistance-R1

In this case, I want to supply a constant current of 50 mA or 0.05A.

So try some new, to find out the R1 as follows:

R1 = 5V/0.05A

= 100 ohms

 

Companents

-7805

-100R Resistor

-0.1uF and 0.33uF Capasitor

-2x1 header